requirements.txt
Django==3.1.1
django-elasticsearch-dsl==7.1.4
django-rest-framework==0.1.0
django-cors-headers==3.5.0
pandas==1.1.3